+++ server/types/mock.go
@@ -68,7 +68,7 @@ func (m *Mock) Validate() error {
 func (m *Mock) Init() {
 	m.State = &MockState{
 		ID:           shortid.MustGenerate(),
-		TimesCount:   &atomic.Int64{},
+		TimesCount:   &atomic.Int32{},
 		CreationDate: time.Now(),
 	}
 
@@ -79,7 +79,7 @@ func (m *Mock) Init() {
 
 func (m *Mock) Verify() bool {
 	isTimesDefined := m.Context.Times > 0
-	hasBeenCalledRightNumberOfTimes := m.State.TimesCount.Load() == int64(m.Context.Times)
+	hasBeenCalledRightNumberOfTimes := m.State.TimesCount.Load() == int32(m.Context.Times)
 	return !isTimesDefined || hasBeenCalledRightNumberOfTimes
 }
 
@@ -91,7 +91,7 @@ func (m *Mock) CloneAndReset() *Mock {
 			ID:           m.State.ID,
 			CreationDate: time.Now(),
 			Locked:       m.State.Locked,
-			TimesCount:   &atomic.Int64{},
+			TimesCount:   &atomic.Int32{},
 		},
 		DynamicResponse: m.DynamicResponse,
 		Proxy:           m.Proxy,
@@ -274,7 +274,7 @@ type MockContext struct {
 
 type MockState struct {
 	ID           string        `json:"id" yaml:"id"`
-	TimesCount   *atomic.Int64 `json:"times_count" yaml:"times_count"`
+	TimesCount   *atomic.Int32 `json:"times_count" yaml:"times_count"`
 	Locked       bool          `json:"locked" yaml:"locked"`
 	CreationDate time.Time     `json:"creation_date" yaml:"creation_date"`
 }
